Move determinant_hashvector to sparse_ci/
