+++ src/aci/aci.h
@@ -42,7 +42,7 @@
 #include "fci/fci_integrals.h"
 #include "mrpt2.h"
 #include "orbital-helpers/unpaired_density.h"
-#include "determinant_hashvector.h"
+#include "sparse_ci/determinant_hashvector.h"
 #include "base_classes/reference.h"
 #include "sparse_ci/sparse_ci_solver.h"
 #include "sparse_ci/determinant.h"
