+++ src/mrci.h
@@ -34,7 +34,7 @@
 
 #include "helpers.h"
 #include "integrals/integrals.h"
-#include "determinant_hashvector.h"
+#include "sparse_ci/determinant_hashvector.h"
 #include "fci/fci_integrals.h"
 #include "sparse_ci/sparse_ci_solver.h"
