+++ src/orbital-helpers/es-nos.h
@@ -37,7 +37,7 @@
 #include "sparse_ci/determinant.h"
 #include "integrals/integrals.h"
 #include "fci/fci_integrals.h"
-#include "determinant_hashvector.h"
+#include "sparse_ci/determinant_hashvector.h"
 #include "operator.h"
 #include "sparse_ci/sparse_ci_solver.h"
