+++ src/sparse_ci/determinant_hashvector.cc
