+++ src/sparse_ci/sorted_string_list.h
@@ -32,7 +32,7 @@
 
 #include "stl_bitset_string.h"
 #include "determinant.h"
-#include "determinant_hashvector.h"
+#include "sparse_ci/determinant_hashvector.h"
 #include "fci/fci_integrals.h"
 #include "sparse_ci/ui64_determinant.h"
