+++ src/sparse_ci/sparse_ci_solver.h
@@ -30,7 +30,7 @@
 #define _sparse_ci_h_
 
 #include "fci/fci_integrals.h"
-#include "determinant_hashvector.h"
+#include "sparse_ci/determinant_hashvector.h"
 #include "operator.h"
 #include "helpers.h"
